using a pointer as an array
refer to char_arr this is one C equivalent. I like to mix C and C++ code together nicely :)
- pointers 
- arrays
- c
- printf()
- chars
remember the mentality of C is "what if everything was a pointer?"

remember that C is extremely permissive and will allow you to do almost anything, another reason to love/hate it.